Fix: 타입 변경, html/js코드 예쁘게

+++ src/main/java/noodlezip/store/dto/MenuRequestDto.java
@@ -16,7 +16,7 @@ public class MenuRequestDto {
     private String menuDescription;
     private String menuImageUrl;
 
-    private Integer ramenCategoryId; // ramenCategory 엔티티의 ID
+    private Long ramenCategoryId; // ramenCategory 엔티티의 ID
     private Integer ramenSoupId;     // ramenSoup 엔티티의 ID
 
     private List<Long> defaultToppingIds; // 기본 토핑 ID 목록 (다대다 관계)
